refactor: move auth services to a separate folder
