+++ src/services/apiMethods/auth/checkUserExist.ts
